backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
